refactor: fixing some imports
