+++ src/modules/models/definitions/content/implementations/ContentModel.ts
@@ -13,7 +13,7 @@ import { IMethodCallModel } from "@models/statements/methodcall/types/IMethodCal
 import { IEventCallModel } from "@models/statements/eventCall/types/IEventCallModel";
 import { ICreateContentDTO } from "../types/ICreateContentDTO";
 import { IIf } from "@models/statements/if/types/IIf";
-import { ITypeName } from "modules/types/ITypeName";
+import { ITypeName } from "@modules/types/ITypeName";
 import { IContinueModel } from "@models/statements/continue/types/IContinueModel";
 import { IReturnModel } from "@models/statements/return/types/IReturnModel";
 import { IAssertModel } from "@models/statements/assert/types/IAssertModel";
@@ -23,11 +23,11 @@ import { IInput } from "@models/definitions/function/types/IInput";
 import { IDoWhileModel } from "@models/statements/dowhile/types/IDoWhileModel";
 import { IDoWhile } from "@models/statements/dowhile/types/IDoWhile";
 import { IForModel } from "@models/statements/for/types/IForModel";
-import { IDataLocation } from "modules/types/IDataLocation";
+import { IDataLocation } from "@modules/types/IDataLocation";
 import { IFor } from "@models/statements/for/types/IFor";
 import { IMappingModel } from "@models/statements/mapping/types/IMappingModel";
-import { IMappingKeyType } from "modules/types/IMappingKeyType";
-import { IMappingTypeName } from "modules/types/IMappingTypeName";
+import { IMappingKeyType } from "@modules/types/IMappingKeyType";
+import { IMappingTypeName } from "@modules/types/IMappingTypeName";
 import { IRequireModel } from "@models/statements/require/types/IRequireModel";
 import { IRevertModel } from "@models/statements/revert/types/IRevertModel";
 import { ICustomErrorcall } from "@models/statements/revert/types/ICustomErrorCall";
