+++ src/modules/models/definitions/stateMapping/implementations/StateMappingModel.ts
@@ -1,6 +1,6 @@
 import { INestedMapping } from "@models/statements/mapping/types/INestedMapping";
-import { IArrayType } from "modules/types/IArrayType";
-import { IElementaryTypeName } from "modules/types/IElementaryTypeName";
+import { IArrayType } from "@modules/types/IArrayType";
+import { IElementaryTypeName } from "@modules/types/IElementaryTypeName";
 import { ICreateStateMappingDTO } from "../types/ICreateStateMappingDTO";
 import { IStateMapping } from "../types/IStateMapping";
 import { IStateMappingModel } from "../types/IStateMappingModel";
