+++ src/modules/models/definitions/stateVariable/types/IStateVariable.ts
@@ -1,5 +1,5 @@
 import { IVariable } from "@models/definitions/stateVariable/types/IVariable";
-import { IVariableStateMutabilityType } from "modules/types/IVariableStateMutabilityType";
+import { IVariableStateMutabilityType } from "@modules/types/IVariableStateMutabilityType";
 
 export interface IStateVariable extends IVariable {
   stateMutability?: IVariableStateMutabilityType;
