+++ src/modules/models/definitions/type/types/ICreateTypeDTO.ts
@@ -1,4 +1,4 @@
-import { IElementaryTypeName } from "modules/types/IElementaryTypeName";
+import { IElementaryTypeName } from "@modules/types/IElementaryTypeName";
 
 export interface ICreateType {
   identifier: string;
