+++ src/modules/models/definitions/using/types/IUsingType.ts
@@ -1,5 +1,5 @@
-import { IArrayType } from "modules/types/IArrayType";
-import { IElementaryTypeName } from "modules/types/IElementaryTypeName";
+import { IArrayType } from "@modules/types/IArrayType";
+import { IElementaryTypeName } from "@modules/types/IElementaryTypeName";
 
 export interface IUsingType {
   regularType?: IElementaryTypeName;
