+++ src/modules/models/statements/mapping/types/ICreateNestedMapping.ts
@@ -1,5 +1,5 @@
-import { IMappingKeyType } from "modules/types/IMappingKeyType";
-import { IMappingTypeName } from "modules/types/IMappingTypeName";
+import { IMappingKeyType } from "@modules/types/IMappingKeyType";
+import { IMappingTypeName } from "@modules/types/IMappingTypeName";
 
 export interface ICreateNestedMapping {
   type: IMappingKeyType;
