+++ src/modules/models/statements/mapping/types/INestedMapping.ts
@@ -1,5 +1,5 @@
-import { IMappingKeyType } from "modules/types/IMappingKeyType";
-import { IMappingTypeName } from "modules/types/IMappingTypeName";
+import { IMappingKeyType } from "@modules/types/IMappingKeyType";
+import { IMappingTypeName } from "@modules/types/IMappingTypeName";
 
 export interface INestedMapping {
   statement: "nested_mapping";
