+++ src/modules/models/statements/variable/types/ICreateVariableDTO.ts
@@ -1,5 +1,5 @@
 import { INewContract } from "@models/statements/newcontract/types/INewContract";
-import { IDataLocation } from "modules/types/IDataLocation";
+import { IDataLocation } from "@modules/types/IDataLocation";
 
 export interface ICreateVariableDTO {
   type: string;
