+++ src/modules/models/statements/variable/types/ILocalVariable.ts
@@ -1,5 +1,5 @@
 import { IVariable } from "@models/definitions/stateVariable/types/IVariable";
-import { IDataLocation } from "modules/types/IDataLocation";
+import { IDataLocation } from "@modules/types/IDataLocation";
 
 export interface ILocalVariable extends IVariable {
   statement: "variable";
