+++ src/modules/models/toplevels/contract/types/IContract.ts
@@ -5,8 +5,8 @@ import { IInput } from "@models/definitions/function/types/IInput";
 import { IOutput } from "@models/definitions/function/types/IOutput";
 import { IInherits } from "@models/toplevels/inherits/types/IInherits";
 import { IStateVariable } from "@models/definitions/stateVariable/types/IStateVariable";
-import { ITypeName } from "modules/types/ITypeName";
-import { IVariableOptions } from "modules/types/IVariableOptions";
+import { ITypeName } from "@modules/types/ITypeName";
+import { IVariableOptions } from "@modules/types/IVariableOptions";
 import { IContractJson } from "./IContractJson";
 import { IFallback } from "@models/definitions/fallback/types/IFallback";
 import { IReceive } from "@models/definitions/receive/types/IReceive";
