+++ src/modules/models/toplevels/contractBody/implementations/ContractBodyModel.ts
@@ -21,11 +21,11 @@ import { IUsingModel } from "@models/definitions/using/types/IUsingModel";
 import { ICreateMappingDTO } from "@models/statements/mapping/types/ICreateMappingDTO";
 import { ICreateVariableDTO } from "@models/statements/variable/types/ICreateVariableDTO";
 import helpers from "@utils/helpers";
-import { IFunctionStateMutabilityType } from "modules/types/IFunctionStateMutabilityType";
-import { IMappingKeyType } from "modules/types/IMappingKeyType";
-import { IMappingTypeName } from "modules/types/IMappingTypeName";
-import { ITypeName } from "modules/types/ITypeName";
-import { IVisibility } from "modules/types/IVisibility";
+import { IFunctionStateMutabilityType } from "@modules/types/IFunctionStateMutabilityType";
+import { IMappingKeyType } from "@modules/types/IMappingKeyType";
+import { IMappingTypeName } from "@modules/types/IMappingTypeName";
+import { ITypeName } from "@modules/types/ITypeName";
+import { IVisibility } from "@modules/types/IVisibility";
 import { inject, injectable } from "tsyringe";
 import { IContractBody } from "../types/IContractBody";
 import { IContractBodyItem } from "../types/IContractBodyItem";
